ibkr_cp: align the TickData with MarketDataResponse obj

+++ packages/core/src/model/trading/quote.rs
@@ -11,7 +11,7 @@ pub struct QuoteRealTimeInfo {
     pub sequence: u64,
     pub timestamp: u64,
     pub current_price: Decimal,
-    pub volume: u64,
+    pub volume: Decimal,
     pub low_price: Option<Decimal>,
     pub high_price: Option<Decimal>,
     pub open_price: Option<Decimal>,
